refactor(core): actionmenu re-structure
