+++ @navikt/core/react/src/action-menu/ActionMenu.types.ts
@@ -1,0 +1,23 @@
+type ActionMenuGroupLabelingProps =
+  | {
+      /**
+       * Adds a visual and accessible label to the group.
+       */
+      label: string;
+      /**
+       * Adds an aria-label to the group.
+       */
+      "aria-label"?: never;
+    }
+  | {
+      /**
+       * Adds an aria-label to the group.
+       */
+      "aria-label": string;
+      /**
+       * Adds a visual and accessible label to the group.
+       */
+      label?: never;
+    };
+
+export type { ActionMenuGroupLabelingProps };
